Backend/requirements.txt
fhirclient==4.1.0
fhirpy==1.3.0
Flask==2.2.2
Flask_APScheduler==1.12.4
Flask_Cors==3.0.10
jmespath==1.0.1
joblib==1.2.0
keras==2.12.0
munch==2.5.0
ndjson==0.3.1
numpy==1.23.2
pandas==1.4.3
pwntools==4.9.0
pytest==7.3.0
python_dateutil==2.8.2
PyYAML==6.0.1
Requests==2.31.0
scikit_learn==1.0.2
tensorflow_macos==2.12.0 # should be tensorflow==2.12.0 in other platforms
tqdm==4.65.0
